+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/IGeneratedEnumTests.java
@@ -2,13 +2,51 @@ package io.github.chrimle.o2jrm.tests;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 import io.github.chrimle.o2jrm.GeneratedSource;
+import io.github.chrimle.o2jrm.models.GeneratedField;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests.IDeprecatedTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests.IEnumTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests.IEnumTests.IConstantsTests;
 import io.github.chrimle.o2jrm.utils.CustomAssertions;
+import java.util.Arrays;
 import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+/**
+ * Tests for <b>all</b> generated {@code enum} classes. Tests are grouped by the <i>method of
+ * configuring</i> the resulting {@code enum}.
+ *
+ * <p>
+ *
+ * <h2>Table of Contents</h2>
+ *
+ * <i>All tests are categorized and grouped accordingly, using {@link Nested} classes. This is the
+ * complete list of tested <i>properties</i> and <i>configurations</i>.<br>
+ *
+ * <p><b>OpenAPI Specification</b>
+ *
+ * <p>{@link IOpenAPITests}
+ *
+ * <ul>
+ *   <li>{@link ISchemaTests `components.schemas.{schema}`}
+ *   <li>{@link IDeprecatedTests `components.schemas.{schema}.deprecated`}
+ *   <li>{@link IEnumTests `components.schemas.{schema}.enum`}
+ *   <li>{@link IConstantsTests `components.schemas.{schema}.enum.{constants}`}
+ * </ul>
+ *
+ * <p><b>`openapi-generator` Configurations</b>
+ *
+ * <p>{@link GeneratorConfigurationTests}
+ *
+ * <ul>
+ *   <li>{@link ConfigOptionsTests configOptions}
+ *   <li>{@link AdditionalEnumTypeAnnotationsTests additionalEnumTypeAnnotations}
+ *   <li>{@link UseEnumCaseInsensitiveTests useEnumCaseInsensitive}
+ * </ul>
+ */
 @DisplayName("Test Generated `enum` classes")
 public abstract class IGeneratedEnumTests {
 
@@ -71,6 +109,81 @@ public abstract class IGeneratedEnumTests {
                 CustomAssertions.assertClassHasMethod(
                     generatedSource.getClassUnderTest(), "getValue"),
                 JsonValue.class);
+          }
+        }
+
+        @Nested
+        @DisplayName("Testing `components.schemas.{schema}.enum.{constants}`")
+        class IConstantsTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generates an `enum` with expected number of constants")
+          void whenEnumHasConstantsThenGeneratedEnumClassHasExpectedNumberOfConstants(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            CustomAssertions.assertClassHasEnumConstants(
+                generatedSource.getClassUnderTest(), generatedSource.generatedFields().length);
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generates `enum` constants with expected names")
+          void whenEnumHasConstantsThenGeneratedEnumClassHasConstantsWithExpectedNames(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            CustomAssertions.assertClassHasEnumConstantsWithNames(
+                generatedSource.getClassUnderTest(),
+                Arrays.stream(generatedSource.generatedFields())
+                    .map(GeneratedField::name)
+                    .toArray(String[]::new));
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "OpenAPI `{schema}.enum.{constants}` -> Generates `enum` constants with expected values")
+          void whenEnumHasConstantsThenGeneratedEnumClassHasConstantsWithExpectedValues(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            CustomAssertions.assertClassHasEnumConstantsWithValues(
+                generatedSource.getClassUnderTest(),
+                Arrays.stream(generatedSource.generatedFields())
+                    .map(GeneratedField::enumValue)
+                    .toArray());
+          }
+        }
+      }
+
+      @Nested
+      @DisplayName("Testing `components.schemas.{schema}.deprecated`")
+      class IDeprecatedTests {
+
+        @Nested
+        @DisplayName("Testing `components.schemas.{schema}.deprecated: false`")
+        class DeprecatedFalseTests {
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generated `enum` class is NOT annotated with `@Deprecated`")
+          void whenEnumIsNotDeprecatedThenGeneratedEnumClassNotIsAnnotatedDeprecated(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.isDeprecated());
+            CustomAssertions.assertClassIsNotAnnotatedWith(
+                generatedSource.getClassUnderTest(), Deprecated.class);
+          }
+        }
+
+        @Nested
+        @DisplayName("Testing `components.schemas.{schema}.deprecated: true`")
+        class DeprecatedTrueTests {
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generated `enum` class is annotated with `@Deprecated`")
+          void whenEnumIsDeprecatedThenGeneratedEnumClassIsAnnotatedDeprecated(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.isDeprecated());
+            CustomAssertions.assertClassIsAnnotatedWith(
+                generatedSource.getClassUnderTest(), Deprecated.class);
           }
         }
       }
